[feat] Add RM_Bitmap class and basic tests

+++ src/rm_rid.h
@@ -15,7 +15,7 @@
 //
 // PageNum: uniquely identifies a page in a file
 //
-typedef int PageNum;
+typedef long long PageNum;
 
 //
 // SlotNum: uniquely identifies a record in a page
